+++ piglegsurgeryweb/uploader/urls.py
@@ -61,4 +61,5 @@ urlpatterns = [
     path("common_review", views.common_review, name="common_review"),
     path("collections/", views.collections, name="collections"),
     path("run_collection/<int:collection_id>/", views.run_collection, name="run_collection"),
+    path("show_collection/<int:collection_id>/", views.show_collection_reports_list, name="show_collection"),
 ]